+++ app/explore/_layout.tsx
@@ -1,0 +1,10 @@
+// import TabLayout from "../(tabs)/_layout";
+
+import CustomTabs from "@/components/CustomTabs";
+
+
+export default function ExploreLayout() {
+    return (
+        <CustomTabs/>
+    );
+}
