+++ app/progress/_layout.tsx
@@ -1,0 +1,7 @@
+import CustomTabs from "@/components/CustomTabs";
+
+export default function ProfileLayout() {
+    return (
+        <CustomTabs />
+    );
+}
