+++ components/CustomTabs.tsx
@@ -1,0 +1,72 @@
+import { Tabs } from 'expo-router';
+import React from 'react';
+import { Platform, View } from 'react-native';
+
+import { HapticTab } from '@/components/HapticTab';
+import { IconSymbol } from '@/components/ui/IconSymbol';
+import TabBarBackground from '@/components/ui/TabBarBackground';
+import { AppColors, Colors } from '@/constants/Colors';
+import { useColorScheme } from '@/hooks/useColorScheme';
+
+const CustomTabs = () => {
+    const colorScheme = useColorScheme();
+
+    return (
+        <Tabs
+            screenOptions={{
+                tabBarActiveTintColor: Colors[colorScheme ?? 'light'].tint,
+                tabBarInactiveTintColor: Colors[colorScheme ?? 'light'].tabIconDefault,
+                headerShown: false,
+                tabBarButton: HapticTab,
+                tabBarBackground: () => (
+                    <View style={{
+                        flex: 1, backgroundColor: AppColors.LightBlue, borderTopColor: '#ddd', borderTopWidth: 0.5
+                    }} />
+                ),
+                tabBarStyle: Platform.select({
+                    ios: {
+                        position: 'absolute',
+                    },
+                    default: {},
+                }),
+            }}>
+            <Tabs.Screen
+                name="progress"
+                options={{
+                    title: '',
+                    tabBarIcon: ({ color }) => <IconSymbol size={28} name="flag.fill" color={color} />,
+                }}
+            />
+            <Tabs.Screen
+                name="explore"
+                options={{
+                    title: '',
+                    tabBarIcon: ({ color }) => <IconSymbol size={28} name='globe.americas.fill' color={color} />,
+                }}
+            />
+            <Tabs.Screen
+                name="index"
+                options={{
+                    title: '',
+                    tabBarIcon: ({ color }) => <IconSymbol size={28} name="house.fill" color={color} />,
+                }}
+            />
+            <Tabs.Screen
+                name="messages"
+                options={{
+                    title: "",
+                    tabBarIcon: ({ color }) => <IconSymbol size={28} name='message.fill' color={color} />,
+                }}
+            />
+            <Tabs.Screen
+                name="profile"
+                options={{
+                    title: '',
+                    tabBarIcon: ({ color }) => <IconSymbol size={28} name="person.circle.fill" color={color} />,
+                }}
+            />
+        </Tabs>
+    );
+};
+
+export default CustomTabs;
